Resolve comments.
